consumer/requirements.txt
kafka-python-ng

